More tweaks for libtiff 4.5.0

+++ tiff/libtiff/tif_hash_set.c
@@ -92,7 +92,7 @@ static const int anPrimes[] = {
 
 static unsigned long TIFFHashSetHashPointer(const void *elt)
 {
-    return (unsigned long)(uintptr_t)((void *)(elt));
+    return (unsigned long)((void *)(elt));
 }
 
 /************************************************************************/
